[ALPS05934067] mrdump: Extend sctlr_elX to 64-bit
Extend sctlr_elX to 64-bit for ARM new spec

MTK-Commit-Id: 6310f8754f8d4d22db4ae6a0aa5eaebd42ba7433

CR-Id: ALPS05934067
Feature: Memory RAM Dump (MRDUMP)

+++ drivers/misc/mediatek/include/mt-plat/mrdump.h
@@ -45,9 +45,9 @@ struct arm32_ctrl_regs {
 };
 
 struct aarch64_ctrl_regs {
-	uint32_t sctlr_el1;
-	uint32_t sctlr_el2;
-	uint32_t sctlr_el3;
+	uint64_t sctlr_el1;
+	uint64_t sctlr_el2;
+	uint64_t sctlr_el3;
 
 	uint64_t tcr_el1;
 	uint64_t tcr_el2;
